Refactored React file structure
