+++ client/src/components/Analytics/SimpleTable.tsx
@@ -1,16 +1,16 @@
-import * as React from 'react';
+import React, { FC } from 'react';
 import {
   Link, Switch, Route, useRouteMatch,
 } from 'react-router-dom';
 import Table from '@material-ui/core/Table';
 import { makeStyles } from '@material-ui/core/styles';
-import TableBody from '@material-ui/core/TableBody';
+import { TableBody } from '@material-ui/core';
 import TableCell from '@material-ui/core/TableCell';
 import TableContainer from '@material-ui/core/TableContainer';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import Paper from '@material-ui/core/Paper';
-import '../styles/application.scss';
+import '../../styles/application.scss';
 
 const useStyles = makeStyles({
   table: {
@@ -25,7 +25,7 @@ const useStyles = makeStyles({
   }
 
 
-function SimpleTable() {
+const SimpleTable: FC = () => {
   const classes = useStyles();
 
   return (
@@ -52,7 +52,7 @@ function SimpleTable() {
       </Table>
     </TableContainer>
   );
-}
+};
 
 export default SimpleTable;
