+++ client/src/components/Analytics/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import Analytics from './Analytics';
+
+export default Analytics;
