+++ client/src/components/App.tsx
@@ -3,22 +3,20 @@ import {
   BrowserRouter as Router, Switch, Route, Link,
 } from 'react-router-dom';
 
+import Home from './Home';
 import '../styles/application.scss';
-import PlaygroundDisplay from './PlaygroundDisplay';
 import Analytics from './Analytics';
 
 const App: FC = () => (
   <div className="app">
     <Router>
-      <Route path="/a">
-        <Analytics />
-      </Route>
       <Switch>
         <Route path="/" exact>
-          <Link to="/a">Analytics</Link>
-          <PlaygroundDisplay />
+          <Home />
         </Route>
-
+        <Route path="/analytics">
+          <Analytics />
+        </Route>
       </Switch>
     </Router>
   </div>
