+++ client/src/components/Home/PlaygroundDisplay.tsx
