+++ client/src/components/Home/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import Home from './Home';
+
+export default Home;
