+++ client/src/components/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+import App from './App';
+
+export default App;
