+++ client/src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import App from './components/App';
+import App from './components';
 
 render(<App />, document.getElementById('root'));
